feat(addon): wire defaultColorFormat through the preview; drop the global color-format picker
Adds defaultColorFormat to the addon's virtual module and live ProjectSnapshot
so a real Storybook/docs preview honors Config.defaultColorFormat end-to-end
(B1/B2 only covered hand-built snapshots). Removes the swatchbookColorFormat
toolbar global and its footer control from both switcher mounts; blocks fall
through to the snapshot default. ColorFormatSelector moves from the addon to
@unpunnyfuns/swatchbook-switcher, still exported for hosts that want their
own footer control.
