+++ .changeset/addon-drop-color-format-picker.md
@@ -1,0 +1,6 @@
+---
+'@unpunnyfuns/swatchbook-addon': minor
+'@unpunnyfuns/swatchbook-switcher': minor
+---
+
+Wire `Config.defaultColorFormat` through the addon's live preview and drop the global color-format picker from the Storybook toolbar; `ColorFormatSelector` moves to `@unpunnyfuns/swatchbook-switcher` for hosts that want their own footer control.
