+++ apps/docs/src/content/docs/reference/addon.mdx
@@ -51,18 +51,17 @@ export default definePreview({
 ## What it registers
 
 - **Preview decorator**: reads the active axis tuple, writes one `data-<prefix>-<axis>="<context>"` attribute per axis on `<html>` + story wrapper, mounts the per-theme CSS, emits `INIT_EVENT` to the manager. The prefix follows `cssVarPrefix` (default `swatch`).
-- **Toolbar tool**: a single swatchbook `IconButton`. Clicking opens a popover containing preset pills, one dropdown per axis, and a color-format picker (`hex` / `rgb` / `hsl` / `oklch` / `raw`). Escape and outside-click close it.
+- **Toolbar tool**: a single swatchbook `IconButton`. Clicking opens a popover containing preset pills and one dropdown per axis. Escape and outside-click close it.
 
 For a browsable tree + diagnostics surface, compose `<Diagnostics />` + `<TokenNavigator />` + `<TokenTable />` on an MDX page. See the [authoring guide](/guides/authoring-doc-stories)'s dashboard example. The [Diagnostics reference](/reference/diagnostics) catalogs every group and message that can appear in the Design Tokens panel.
 
 ## Globals
 
-| Key                     | Type                                          | Purpose                                                                                                                                                           |
-| ----------------------- | --------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `swatchbookAxes`        | `Record<string, string>`                      | Active tuple. Preferred input.                                                                                                                                    |
-| `swatchbookColorFormat` | `'hex' \| 'rgb' \| 'hsl' \| 'oklch' \| 'raw'` | Display-only color format consumed by the blocks (`TokenTable`, `ColorPalette`, `TokenDetail`, `TokenNavigator`). Does **not** affect emitted CSS. Default `hex`. |
+| Key              | Type                     | Purpose                        |
+| ---------------- | ------------------------ | ------------------------------ |
+| `swatchbookAxes` | `Record<string, string>` | Active tuple. Preferred input. |
 
-The toolbar writes `swatchbookAxes`. `swatchbookColorFormat` is toggled from the color-format picker in the popover; `hex` falls back to space-separated `rgb()` (flagged with a ⚠ warning marker) for colors outside sRGB gamut.
+The toolbar writes `swatchbookAxes`. Blocks' starting color format comes from [`Config.defaultColorFormat`](/reference/config#defaultcolorformat) instead of a toolbar control; `hex` falls back to space-separated `rgb()` (flagged with a ⚠ warning marker) for colors outside sRGB gamut. See [`useColorFormat`](/reference/blocks/hooks#usecolorformat) for the full precedence order.
 
 ## Per-story overrides
 
@@ -106,13 +105,7 @@ The block-side hooks (`useSwatchbookData`, `useActiveTheme`, `useActiveAxes`, `u
 ## Exported constants
 
 ```ts
-import {
-  ADDON_ID,
-  AXES_GLOBAL_KEY,
-  COLOR_FORMAT_GLOBAL_KEY,
-  TOOL_ID,
-  VIRTUAL_MODULE_ID,
-} from '@unpunnyfuns/swatchbook-addon';
+import { ADDON_ID, AXES_GLOBAL_KEY, TOOL_ID, VIRTUAL_MODULE_ID } from '@unpunnyfuns/swatchbook-addon';
 ```
 
 Useful when wiring the addon into custom tooling or writing interaction tests that read the globals directly.
