+++ apps/storybook/src/stories/Toolbar.stories.tsx
@@ -9,9 +9,10 @@
  * contracts: axis switch, preset apply, footer-slot integration, and
  * the "modified-since-preset" reset path.
  *
- * `<ColorFormatSelector>` — the real footer the addon slots into the
- * switcher in production — has its own dedicated component tests in
- * `packages/addon/test/color-format-selector.browser.test.tsx`; the
+ * `<ColorFormatSelector>` — exported from `@unpunnyfuns/swatchbook-switcher`
+ * for hosts that want a footer control, though neither switcher mount
+ * renders it by default — has its own dedicated component tests in
+ * `packages/switcher/test/color-format-selector.browser.test.tsx`; the
  * footer-slot story here uses a stand-in button so the integration
  * isn't coupled to the selector's specifics.
  *
@@ -170,10 +171,11 @@ export const SwitchPreset = meta.story({
 });
 
 /**
- * Footer slot integration — host content (the addon uses
- * `<ColorFormatSelector>` here; the story uses a stand-in button to
- * decouple from the selector's own tests in `packages/addon/test/`)
- * renders alongside the axis rows and receives clicks normally.
+ * Footer slot integration — host content (a consumer could slot
+ * `@unpunnyfuns/swatchbook-switcher`'s `<ColorFormatSelector>` here; the
+ * story uses a stand-in button to decouple from the selector's own tests
+ * in `packages/switcher/test/`) renders alongside the axis rows and
+ * receives clicks normally.
  */
 export const FooterSlot = meta.story({
   play: async ({ canvasElement }) => {
