--- packages/addon/src/ColorFormatSelector.tsx
@@ -1,79 +1,0 @@
-import { COLOR_FORMATS } from '@unpunnyfuns/swatchbook-core/color-formats';
-import type { ColorFormat } from '@unpunnyfuns/swatchbook-core/color-formats';
-import React from 'react';
-import type { ReactElement } from 'react';
-
-export { COLOR_FORMATS };
-export type { ColorFormat };
-
-// Storybook-addon-specific pill row for picking how color sub-values
-// render in swatchbook blocks (hex / rgb / hsl / oklch / raw). Lives in
-// the addon rather than the shared switcher because color format is a
-// blocks-rendering concern, not a theming one — the docs-site navbar
-// switcher has no consumer for it.
-//
-// Reuses the `sb-switcher__*` class names so styling stays consistent
-// with the rest of the toolbar popover. The switcher's CSS is already
-// loaded on the page because this selector only renders inside
-// `<ThemeSwitcher>`'s `footer` prop.
-//
-// Uses `React.createElement` (via `h`) to survive embedding in Storybook's
-// manager bundle, which doesn't expose `react/jsx-runtime`.
-
-// COLOR_FORMATS comes from core's colorjs-free `color-formats` leaf, not
-// from blocks: the manager bundle can't pull `@unpunnyfuns/swatchbook-blocks`
-// (its barrel has a top-level `import 'virtual:swatchbook/tokens'` the
-// manager has no resolver for), and core's leaf carries no colorjs weight.
-// Exported for the parity test that guards this hand-listed label set
-// against drifting from core's canonical `COLOR_FORMATS`.
-export const COLOR_FORMAT_OPTIONS: readonly { id: ColorFormat; label: string }[] = [
-  { id: 'hex', label: 'Hex' },
-  { id: 'rgb', label: 'RGB' },
-  { id: 'hsl', label: 'HSL' },
-  { id: 'oklch', label: 'OKLCH' },
-  { id: 'raw', label: 'Raw (JSON)' },
-];
-
-const h = React.createElement;
-
-interface ColorFormatSelectorProps {
-  active: ColorFormat;
-  onSelect(next: ColorFormat): void;
-}
-
-export function ColorFormatSelector({ active, onSelect }: ColorFormatSelectorProps): ReactElement {
-  return h(
-    'div',
-    null,
-    h(
-      'div',
-      { className: 'sb-switcher__section-label', id: 'sb-color-format-label' },
-      'Color format',
-    ),
-    h(
-      'div',
-      {
-        className: 'sb-switcher__section-body',
-        role: 'group',
-        'aria-labelledby': 'sb-color-format-label',
-      },
-      ...COLOR_FORMAT_OPTIONS.map((opt) => {
-        const isActive = opt.id === active;
-        return h(
-          'button',
-          {
-            key: `color-format/${opt.id}`,
-            type: 'button',
-            'aria-pressed': isActive,
-            'aria-label': `${opt.label} color format`,
-            onClick: () => onSelect(opt.id),
-            className: isActive
-              ? 'sb-switcher__pill sb-switcher__pill--active'
-              : 'sb-switcher__pill',
-          },
-          opt.label,
-        );
-      }),
-    ),
-  );
-}
