+++ packages/addon/test/virtual-stub.ts
@@ -22,6 +22,7 @@ export const presets = [] as const;
 export const diagnostics = [] as const;
 export const css = '';
 export const cssVarPrefix = 'sb';
+export const defaultColorFormat = 'hex';
 export const listing = {};
 export const defaultTuple: Record<string, string> = { mode: 'Light' };
 export const tokenGraph: TokenGraph = {
