+++ packages/switcher/src/index.ts
@@ -1,3 +1,5 @@
 export { ThemeSwitcher, type ThemeSwitcherProps } from '#/ThemeSwitcher.tsx';
 export { presetTuple } from '#/preset-tuple.ts';
-export type { SwitcherAxis, SwitcherPreset } from '#/types.ts';
+export { ColorFormatSelector, COLOR_FORMAT_OPTIONS } from '#/ColorFormatSelector.tsx';
+export { COLOR_FORMATS } from '#/types.ts';
+export type { ColorFormat, SwitcherAxis, SwitcherPreset } from '#/types.ts';
